Started to move share-related functions into new Share and RemoteShare classes.
Fixed stupid bug in File.write which was setting the length method rather
than the _length attribute.

Investigated problems with transfer_file which were partly imaginary: added
check for the position in the file received with data in put operations so
that, only when the position is as expected, is the data written to the
File object.

The pput method was not increasing the start position within the file for
referencing successive blocks of data, causing the transfer to get stuck
after a block of data.