Reverted permissions checked for reading and writing by remote clients to the "other" permissions on files. This means that we don't have to check who the owner is before trying to determine whether the permissions allow reading or writing.
Changed the Messages object so that it uses a dictionary internally to
organise messages. When our client expects messages using a certain ID,
it asks the Messages object to add an entry for this ID in its internal
dictionary; reply-type messages ("D", "R", "S", "E", "d", "r", "w") are
discarded if they do not have a corresponding entry in this dictionary.
When no more messages using the ID are expected then the client asks the
Messages object to remove the corresponding dictionary entry for that ID.