Updated version number and fixed future-proofed RISC OS date handling.
Peer.send_file:

Changed D message so that the offset within a current block being
transferred is returned to the remote host. Before, it was assumed that
the absolute offset was used.

Changed R message at the end of a block so that the absolute value of the
end offset is given. Before, the offset relative to the start of the block
was given. Additionally, the length of data transferred in the block is
also sent.