Ignore share broadcasts from our own host.
A race condition when starting up caused the code receiving share broadcasts
to see our share broadcast before our share was added to the shares dictionary.
This would cause the share to be added to the dictionary as a RemoteShare
instead of a Share.  access.py would fail to shut down as it tried to signal
the RemoteShare's broadcast thread to stop instead of the Share's broadcast
thread.